backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
groq==0.12.0
python-dotenv==1.0.1
httpx==0.27.2
